requirements.txt
Flask==2.0.2
Flask-RESTful==0.3.9
tensorflow==2.12.0
numpy==1.23.5
gunicorn
google-cloud-storage==2.9.0
#python image processing library
Pillow==8.3.2
matplotlib